componetizacion de la landing page update of demo update webapp info page
