+++ frontend/landing-page/nginx.conf
@@ -16,6 +16,8 @@ server {
     image/jpeg jpg jpeg;
     image/gif gif;
     image/webp webp;
+    video/webm webm;
+    video/mp4 mp4;
     font/woff2 woff2;
     application/xml xml;
     text/plain txt;
@@ -37,7 +39,7 @@ server {
   }
 
   # Static assets: long cache
-  location ~* \.(js|css|svg|png|jpg|jpeg|gif|webp|woff2|ico)$ {
+  location ~* \.(js|css|svg|png|jpg|jpeg|gif|webp|woff2|ico|webm|mp4)$ {
     expires 1y;
     add_header Cache-Control "public, immutable";
   }
